Reworked the Website Landing Page to not be so reliant upon flexboxes. Made use of the "margin-inline: auto" function to make the elements centered without the need for flexbox wisardry.
Made a simple navigation bar, It will need to be made web Responsive.

Added a canvas element in the header and worked with the css to make it so that the H1 element would be renderd above the Canvas Element.
Asked ChatGPT to make a bubble JS Script to ensure that the canvas was working as intended. It will be reworked without the AI at another time.